fix(sanction): CodeRabbit·lim-haeun 리뷰 2차 반영 — 8개 항목 수정
- Report.create(): reportedUserId 파라미터 추가 (제재 누적 집계 버그 수정)
- ReportService: create() 호출 시 reportedUserId 파생 후 저장, 이벤트 발행 블록 헬퍼 메서드로 추출
- ReportRepository/ReportAcceptedEvent/SanctionService/SanctionType: acceptedCount int → long
- AccountRepository: findExpiredSystemSanctions JPQL enum FQCN → @Param 파라미터화
- SuspendExpiryScheduler: forEach → 항목별 try-catch, 로그에 성공 건수 추가
- AdminSanctionService: TOCTOU recheck 추가, ifPresent → ifPresentOrElse + warn 로그
- UserSanctionRepository: WARNING 만료 처리 Javadoc 명확화

+++ src/main/java/com/chunbaetour/domain/report/entity/SanctionType.java
@@ -22,8 +22,8 @@ public enum SanctionType {
         return this.severity > other.severity;
     }
 
-    /** 누적 승인 건수 → 제재 단계 매핑 (임계치: 3/5/10/15). */
-    public static SanctionType fromCount(int count) {
+    /** 누적 승인 건수 → 제재 단계 매핑 (임계치: 3 이상/5 이상/10 이상/15 이상). */
+    public static SanctionType fromCount(long count) {
         if (count >= 15) return PERMANENT;
         if (count >= 10) return SUSPEND_30D;
         if (count >= 5)  return SUSPEND_7D;
